Update CameraPreview.java
Changed behavior so that asking for permissions is treated as a failure, even if the permission is granted. The reason for this change is that on modern Android devices, even if permission is granted, the camera resource is not immediately available, and the plugin call ends up failing anyways. Except, if we don't fail it here, the camera preview surface will still render (without proper access to the camera resource) and if the user taps it, the application will crash out.